@@ -56,7 +56,7 @@
         "@types/jsonfile": "6.0.1",
         "@types/jwt-decode": "2.2.1",
         "@types/lodash": "4.14.161",
-        "@types/mocha": "8.2.1",
+        "@types/mocha": "^8.2.3",
         "@types/node": "^14.18.9",
         "@types/node-notifier": "8.0.1",
         "@types/qs": "6.9.3",
@@ -87,14 +87,14 @@
         "ora": "5.4.0",
         "prettier": "2.4.1",
         "reflect-metadata": "0.1.13",
-        "ts-node": "10.4.0",
+        "ts-node": "^10.9.1",
         "tsconfig-paths": "3.11.0",
-        "ttypescript": "1.5.12",
-        "typescript": "4.5.2",
+        "typescript": "^4.7.4",
         "typescript-transform-paths": "2.2.4",
         "wdio-chromedriver-service": "^7.3.2",
         "wdio-reportportal-reporter": "^6.5.0",
-        "wdio-reportportal-service": "^6.4.0"
+        "wdio-reportportal-service": "^6.4.0",
+        "wdio-wait-for": "^2.2.6"
     },
     "dependencies": {
         "@bbc/cucumber-slicer": "1.0.9",
@@ -106,6 +106,7 @@
         "@wdio/devtools-service": "7.20.9",
         "@wdio/sauce-service": "^7.16.13",
         "@wdio/selenium-standalone-service": "7.20.8",
+        "@wdio/mocha-framework": "^7.20.7",
         "ajv": "^6.12.6",
         "allure-commandline": "2.13.0",
         "allure-cucumberjs": "2.0.0-beta.9",
