@@ -2,12 +2,12 @@
 * main page object containing all methods, selectors and functionality
 * that is shared across all page objects
 */
-module.exports = class Page {
+export default class Page {
     /**
     * Opens a sub page of the page
     * @param path path of the sub page (e.g. /path/to/page.html)
     */
-    open (path) {
+    public open (path: string) {
         return browser.url(`/${path}`);
     }
 };
