@@ -8,6 +8,6 @@ export default class Page {
     * @param path path of the sub page (e.g. /path/to/page.html)
     */
     open (path: string) {
-        return browser.url(`https://the-internet.herokuapp.com/${path}`)
+        return browser.url(`/${path}`);
     }
 }
